Move test out
