+++ packages/wonder-blocks-form/generated-snapshot.test.js
@@ -73,74 +73,4 @@ describe("wonder-blocks-form", () => {
         const tree = renderer.create(example).toJSON();
         expect(tree).toMatchSnapshot();
     });
-    it("example 2", () => {
-        const {View} = require("@khanacademy/wonder-blocks-core");
-        const {StyleSheet, css} = require("aphrodite");
-
-        const styles = StyleSheet.create({
-            cell: {
-                padding: 10,
-            },
-            checkbox: {
-                marginLeft: 5,
-                marginRight: 5,
-            },
-        });
-
-        const states = ["default", "error", "disabled"];
-        const clickableStates = ["default", "hovered", "pressed"];
-        const checkedStates = [false, true];
-
-        const example = (
-            <table>
-                <thead>
-                    <tr>
-                        <th />
-                        <th colSpan="2">Default</th>
-                        <th colSpan="2">Hover/focus</th>
-                        <th colSpan="2">Active/press</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    {states.map((state) => (
-                        <tr key={state}>
-                            <td>{state}</td>
-                            {clickableStates.map((clickableState) => {
-                                const checkboxes = checkedStates.map(
-                                    (checked) => {
-                                        return (
-                                            <td
-                                                key={`${clickableState}-${checked}`}
-                                                className={css(styles.cell)}
-                                            >
-                                                <CheckboxCore
-                                                    checked={checked}
-                                                    disabled={
-                                                        state === "disabled"
-                                                    }
-                                                    error={state === "error"}
-                                                    hovered={
-                                                        clickableState ===
-                                                        "hovered"
-                                                    }
-                                                    pressed={
-                                                        clickableState ===
-                                                        "pressed"
-                                                    }
-                                                    style={[styles.checkbox]}
-                                                />
-                                            </td>
-                                        );
-                                    },
-                                );
-                                return checkboxes;
-                            })}
-                        </tr>
-                    ))}
-                </tbody>
-            </table>
-        );
-        const tree = renderer.create(example).toJSON();
-        expect(tree).toMatchSnapshot();
-    });
 });
